+++ cmd/github-actions-adapter/serve/serve_test.go
@@ -197,7 +197,7 @@ func (p *MockRadiclePatch) Comment(ctx context.Context, repoID, patchID, revisio
 	}
 
 	if !strings.Contains(message, "Checking") {
-		if totalWorkflows != strings.Count(message, "[") {
+		if totalWorkflows != strings.Count(message, "[")/2 {
 			p.t.Error("total workflows do not match message")
 			return errors.New("total workflows do not match message")
 		}
@@ -455,7 +455,9 @@ func TestGitHubActions_PreparePatchCommentMessage(t *testing.T) {
 					{WorkflowID: "2", WorkflowName: "UnitTests", WorkflowResult: githubops.WorkflowResultFailure},
 				},
 			},
-			expected: "### GitHub Actions Result: success ✅\n Workflows:\n - [BuildTest (1) ✅](https://github.com/testUser/testRepo/actions/runs/1 \"success\")\n - [UnitTests (2) ❌](https://github.com/testUser/testRepo/actions/runs/2 \"failure\")",
+			expected: "GitHub Actions Result: success ✅  \n Workflows:  \n " +
+				"- BuildTest ([#1](https://github.com/testUser/testRepo/actions/runs/1)) [✅](# \"success\")  \n " +
+				"- UnitTests ([#2](https://github.com/testUser/testRepo/actions/runs/2)) [❌](# \"failure\")",
 		},
 		{
 			name: "PreparePatchCommentMessage is successful using only failed results",
@@ -466,7 +468,9 @@ func TestGitHubActions_PreparePatchCommentMessage(t *testing.T) {
 					{WorkflowID: "2", WorkflowName: "UnitTests", WorkflowResult: githubops.WorkflowResultFailure},
 				},
 			},
-			expected: "### GitHub Actions Result: failure ❌\n Workflows:\n - [BuildTest (1) ✅](https://github.com/testUser/testRepo/actions/runs/1 \"success\")\n - [UnitTests (2) ❌](https://github.com/testUser/testRepo/actions/runs/2 \"failure\")",
+			expected: "GitHub Actions Result: failure ❌  \n Workflows:  \n " +
+				"- BuildTest ([#1](https://github.com/testUser/testRepo/actions/runs/1)) [✅](# \"success\")  \n " +
+				"- UnitTests ([#2](https://github.com/testUser/testRepo/actions/runs/2)) [❌](# \"failure\")",
 		},
 		{
 			name: "PreparePatchCommentMessage is successful using mixed results",
@@ -478,7 +482,10 @@ func TestGitHubActions_PreparePatchCommentMessage(t *testing.T) {
 					{WorkflowID: "3", WorkflowName: "IntegrationTests", WorkflowResult: "otherResult"},
 				},
 			},
-			expected: "### GitHub Actions Result: failure ❌\n Workflows:\n - [BuildTest (1) ✅](https://github.com/testUser/testRepo/actions/runs/1 \"success\")\n - [UnitTests (2) ❌](https://github.com/testUser/testRepo/actions/runs/2 \"failure\")\n - [IntegrationTests (3) ⚠️️](https://github.com/testUser/testRepo/actions/runs/3 \"otherResult\")",
+			expected: "GitHub Actions Result: failure ❌  \n Workflows:  \n - " +
+				"BuildTest ([#1](https://github.com/testUser/testRepo/actions/runs/1)) [✅](# \"success\")  \n - " +
+				"UnitTests ([#2](https://github.com/testUser/testRepo/actions/runs/2)) [❌](# \"failure\")  \n - " +
+				"IntegrationTests ([#3](https://github.com/testUser/testRepo/actions/runs/3)) [⚠️️](# \"otherResult\")",
 		},
 	}